+++ consensus/ibft/currentstate/state.go
@@ -18,7 +18,8 @@ const (
 	AcceptState IbftState = iota
 	RoundChangeState
 	ValidateState // including prepare, commit, and post commit state stage
-	CommitState
+	CommitState   // committed
+	FinState      // finish current sequence
 )
 
 // String returns the string representation of the passed in state
@@ -35,12 +36,17 @@ func (i IbftState) String() string {
 
 	case CommitState:
 		return "CommitState"
+
+	case FinState:
+		return "FinState"
 	}
 
 	panic(fmt.Sprintf("BUG: Ibft state not found %d", i))
 }
 
 // CurrentState defines the current state object in IBFT
+//
+// NOTE: not thread safe, better to wrap it in mutext for concurrent use case
 type CurrentState struct {
 	// validators represent the current validator set
 	validators validator.Validators
@@ -56,6 +62,9 @@ type CurrentState struct {
 
 	// Current view
 	view *proto.View
+
+	// additionalTimeout is for block building, which might consumes longer than we thought
+	additionalTimeout time.Duration
 
 	// List of prepared messages
 	prepared map[types.Address]*proto.MessageReq
@@ -82,6 +91,21 @@ func NewState() *CurrentState {
 	c.ResetRoundMsgs()
 
 	return c
+}
+
+func (c *CurrentState) Clear(height uint64) {
+	c.block = nil
+	c.proposer = types.ZeroAddress
+	c.view = &proto.View{
+		Sequence: height,
+		Round:    0,
+	}
+	c.err = nil
+	c.prepared = map[types.Address]*proto.MessageReq{}
+	c.committed = map[types.Address]*proto.MessageReq{}
+	c.canonicalSeal = nil // reset canonical seal when round change
+	c.roundMessages = map[uint64]map[types.Address]*proto.MessageReq{}
+	c.locked = false
 }
 
 // GetState returns the current state
@@ -199,7 +223,8 @@ func (c *CurrentState) MessageTimeout() time.Duration {
 		timeout += time.Duration(int64(math.Pow(2, float64(c.Round())))) * time.Second
 	}
 
-	return timeout
+	// add more time for long time block producting
+	return timeout + c.additionalTimeout
 }
 
 // ResetRoundMsgs resets the prepared, committed and round messages in the current state
@@ -241,6 +266,11 @@ func (c *CurrentState) SetView(view *proto.View) {
 
 func (c *CurrentState) View() *proto.View {
 	return c.view
+}
+
+// SetAdditionalTimeout sets an additional timeout for potentially long builds
+func (c *CurrentState) SetAdditionalTimeout(d time.Duration) {
+	c.additionalTimeout = d
 }
 
 func (c *CurrentState) IsLocked() bool {
